Crate pallet_mmr_rpc

Node-specific RPC methods for interaction with Merkle Mountain Range pallet.

Modules

gen_client

The generated client module.

Structs

LeafProof

Retrieved MMR leaf and its proof.

Mmr

An implementation of MMR specific RPC methods.

MmrApiClient

The Client.

Traits

MmrApi

MMR RPC methods.

MmrRuntimeApi

API to interact with MMR pallet.

Type Definitions

LeafIndex

A type to describe leaf position in the MMR.